Bluetooth shutdown is a racy mess
bluetooth::stop_service() and bluetooth::exit() are being executed concurrently on kodi
shutdown if the user is in the Bluetooth window during shutdown. Both functions are
checking for, and then deleting, self.discovery_thread.

Since only one of the function calls will successfully delete self.discovery_thread, the
other function call will throw an AttributeError exception. The problem is that the
check-then-delete pattern is not thread safe - the object can be deleted in the first
thread after it has been checked (but before it is deleted) in the second thread.

Ignoring the exception is a hack until someone cares to fix it properly.